--- kramer/src/main/java/com/chiralbehaviors/layout/primitives/LabelCell.java
@@ -1,57 +1,0 @@
-/**
- * Copyright (c) 2017 Chiral Behaviors, LLC, all rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.chiralbehaviors.layout.primitives;
-
-import com.chiralbehaviors.layout.PrimitiveLayout;
-import com.chiralbehaviors.layout.cell.LayoutCell;
-import com.chiralbehaviors.layout.schema.SchemaNode;
-import com.fasterxml.jackson.databind.JsonNode;
-
-import javafx.scene.control.Label;
-
-/**
- * @author halhildebrand
- *
- */
-public class LabelCell implements LayoutCell<Label> {
-    private static final String DEFAULT_STYLE = "primitive";
-    private final Label         label;
-
-    public LabelCell(PrimitiveLayout layout) {
-        label = new Label();
-        initialize(layout.getField());
-        label.setWrapText(true);
-        initialize(DEFAULT_STYLE);
-    }
-
-    @Override
-    public Label getNode() {
-        return label;
-    }
-
-    @Override
-    public boolean isReusable() {
-        return true;
-    }
-
-    @Override
-    public void updateItem(JsonNode item) {
-        label.setText(SchemaNode.asText(item));
-        getNode().pseudoClassStateChanged(PSEUDO_CLASS_FILLED, item != null);
-    }
-
-}
